factor out code detecting missing NCBI model
also, if the model is missing, throw a custom error

Previous commits in this branch made it possible for users
to run data_translators/genome_annotation_translator.py
without triggering the code that tells them how to download
the NCBI model. This should fix that.

Throwing a custom error will also allow users of this code to
handle different failures modes (maybe ENSEMBL is publishing
a non-standard GFF3 file) in different ways.